Add retry to monitoring while on replication.
This is a workaround.
If PostgreSQL has all connection slots used, it will refuse a new connection,
and the resource-agent will consider this to be a failure, and pacemaker will
try to restart PostgreSQL on master, sometimes it fails, and causes a failover.

Recovering from a failover is a costly task while using streaming replication
and large databases.

The root problem needs further investigation, this is a workaround while the
*actual* problem is found and fixed.